debugging

+++ src/ultrasonic/src/heartbeat_subscriber.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 import rospy
 import time
-from std_msgs.msg import Float32
+from std_msgs.msg import Float64
 
 def callback(data):     
     time_elapsed = time.time() - data.data
@@ -11,7 +11,7 @@ def callback(data):
 def listener(): 
     rospy.init_node('ultrasonic_stethoscope', anonymous=True)
 
-    rospy.Subscriber("ultrasonic_heartbeat", Float32, callback)
+    rospy.Subscriber("ultrasonic_heartbeat", Float64, callback)
 
     # spin() simply keeps python from exiting until this node is stopped
     rospy.spin()
